+++ tests/framework/db/sqlite/provider/CommandProvider.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace yiiunit\framework\db\sqlite\provider;
+
+final class CommandProvider extends \yiiunit\framework\db\provider\AbstractCommandProvider
+{
+    protected static string $driverName = 'sqlite';
+}